Load execute_pipeline Lua script from file.
